+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+# 0.11.3
+ * Fix bug in `c:Membrane.Pipeline.handle_call/3` [#526](https://github.com/membraneframework/membrane_core/pull/526).
+
 # 0.11.2
  * Fix bug in Membrane.ChildrenSpec.child/3 spec.